requirements.txt
asgiref==3.8.1
attrs==24.2.0
autoflake==2.3.1
black==24.10.0
click==8.1.7
Django==5.1.1
django-filter==24.3
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.2
drf-spectacular==0.27.2
inflection==0.5.1
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
pyflakes==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
uritemplate==4.1.1
